gemini-html-translator-render-template/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
google-generativeai
tenacity
